+++ frontend/src/pages/SpaceList/index.tsx
@@ -1,4 +1,5 @@
 /**  @jsxImportSource @emotion/react */
+import SpaceCard from '../../components/SpaceCard';
 import PageTitle from '../../components/_common/PageTitle';
 import styles from './styles';
 
@@ -14,19 +15,6 @@ const spaces = [
       'https://images.unsplash.com/photo-1517336714731-489689fd1ca8?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1626&q=80',
   },
 ];
-
-type SpaceCardProps = {
-  spaceName: string;
-  imageURL: string;
-};
-
-const SpaceCard = ({ spaceName, imageURL }: SpaceCardProps) => {
-  return (
-    <div css={styles.spaceCard({ imageURL })}>
-      <span css={styles.cardTitle}>{spaceName}</span>
-    </div>
-  );
-};
 
 const SpaceList = () => {
   return (
